# Food-Delivery/api_gateway_service/docs/swagger.yaml
basePath: /
definitions:
...
  courier.CreateOrderRequest:
    properties:
      customer_id:
        description: Mijozning ID'si
        type: string
      delivery_address:
        description: Yetkazib berish manzili
        type: string
      order_details:
        description: Buyurtma tafsilotlari
        type: string
      payment_status:
        allOf:
        - $ref: '#/definitions/courier.PaymentStatus'
        description: To'lov holati
    type: object
  courier.CreateTaskRequest:
    properties:
      assigned_to:
        type: integer
      description:
        type: string
      due_date:
        description: ISO 8601 date format (e.g., "2023-12-31")
        type: string
      status:
        $ref: '#/definitions/courier.TaskStatus'
      title:
        type: string
    type: object
...
  courier.GetAllTasksResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/courier.TaskResponse'
        type: array
    type: object
  courier.GetCourierOrderHistoryResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/courier.Order'
        type: array
    type: object
  courier.GetRecommendedOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/courier.Order'
        type: array
    type: object
  courier.Order:
    properties:
      customer_id:
        description: Mijozning ID'si
        type: string
      delivery_address:
        description: Yetkazib berish manzili
        type: string
      order_date:
        description: Buyurtma sanasi
        type: string
      order_details:
        description: Buyurtma tafsilotlari
        type: string
      order_id:
        description: Buyurtmaning noyob ID'si
        type: string
      payment_status:
        allOf:
        - $ref: '#/definitions/courier.PaymentStatus'
        description: To'lov holati
    type: object
...
  courier.OrderResponse:
    properties:
      message:
        description: Javob xabari
        type: string
      order:
        allOf:
        - $ref: '#/definitions/courier.Order'
        description: Buyurtma ma'lumotlari (Create va Get operatsiyalari uchun)
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
  courier.OrderStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    type: integer
    x-enum-varnames:
    - OrderStatus_PENDING_Status
    - OrderStatus_ACCEPTED
    - OrderStatus_IN_PROGRESS
    - OrderStatus_DELIVERED
    - OrderStatus_CANCELED
  courier.PaymentStatus:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      PaymentStatus_PAID: 'To''lov holati: to''langan'
      PaymentStatus_PENDING: 'To''lov holati: kutilyapti'
    x-enum-varnames:
    - PaymentStatus_PENDING
    - PaymentStatus_PAID
  courier.TaskResponse:
    properties:
      assigned_to:
        type: integer
      created_at:
        description: ISO 8601 timestamp format (e.g., "2023-12-31T12:34:56Z")
        type: string
      description:
        type: string
      due_date:
        description: ISO 8601 date format (e.g., "2023-12-31")
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/courier.TaskStatus'
      title:
        type: string
      updated_at:
        description: ISO 8601 timestamp format (e.g., "2023-12-31T12:34:56Z")
        type: string
    type: object
  courier.TaskStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - TaskStatus_TODO
    - TaskStatus_IN_PROGRESS
    - TaskStatus_DONE
...
  courier.UpdateOrderRequest:
    properties:
      customer_id:
        description: Mijozning ID'si
        type: string
      delivery_address:
        description: Yetkazib berish manzili
        type: string
      order_details:
        description: Buyurtma tafsilotlari
        type: string
      order_id:
        description: Buyurtmaning noyob ID'si
        type: string
      payment_status:
        allOf:
        - $ref: '#/definitions/courier.PaymentStatus'
        description: To'lov holati
    type: object
  courier.UpdateOrderStatusRequest:
    properties:
      order_id:
        type: string
      status:
        $ref: '#/definitions/courier.OrderStatus'
    type: object
  courier.UpdateOrderStatusResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  courier.UpdateTaskRequest:
    properties:
      assigned_to:
        type: integer
      description:
        type: string
      due_date:
        description: ISO 8601 date format (e.g., "2023-12-31")
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/courier.TaskStatus'
      title:
        type: string
    type: object
...
  product.CardStatus:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - CardStatus_OR_PENDING
    - CardStatus_COMPLETED
    - CardStatus_CANCELED
...
  product.CreateOrderProductRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/product.CardStatus'
        description: Buyurtma holati
      total_price:
        description: Buyurtma umumiy summasi
        type: number
      user_id:
        description: Buyurtma bergan foydalanuvchi ID'si
        type: string
    type: object
...
  product.OrderProduct:
    properties:
      created_at:
        description: Buyurtma yaratilgan vaqti
        type: string
      order_id:
        description: Buyurtma noyob ID'si
        type: string
      status:
        allOf:
        - $ref: '#/definitions/product.CardStatus'
        description: Buyurtma holati
      total_price:
        description: Buyurtma umumiy summasi
        type: number
      user_id:
        description: Buyurtma bergan foydalanuvchi ID'si
        type: string
    type: object
  product.OrderProductListResponse:
    properties:
      orders:
        description: Buyurtmalar ro'yxati
        items:
          $ref: '#/definitions/product.OrderProduct'
        type: array
    type: object
  product.OrderProductResponse:
    properties:
      message:
        description: Javob xabari
        type: string
      order:
        allOf:
        - $ref: '#/definitions/product.OrderProduct'
        description: Buyurtma ma'lumotlari (Create va Get operatsiyalari uchun)
      success:
        description: Amaliyot muvaffaqiyatli bo'lganligini bildiruvchi flag
        type: boolean
    type: object
...
  product.UpdateOrderProductRequest:
    properties:
      order_id:
        description: Buyurtma noyob ID'si
        type: string
      status:
        allOf:
        - $ref: '#/definitions/product.CardStatus'
        description: Buyurtma holati
      total_price:
        description: Buyurtma umumiy summasi
        type: number
    type: object
...
info:
  contact: {}
  description: API for Food Delivery System resources
  title: Food Delivery System API
  version: "1.0"
paths:
...
  /api/orders:
    get:
      description: Retrieve a list of all orders with optional filters
      parameters:
      - description: User ID Filter
        in: query
        name: userId
        type: string
      - description: Order Status Filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.OrderProductListResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List all orders
      tags:
      - Orders
    post:
      consumes:
      - application/json
      description: Add a new order to the system
      parameters:
      - description: Order Data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/product.CreateOrderProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/product.OrderProductResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Modify details of an existing order
      parameters:
      - description: Order Update Data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/product.UpdateOrderProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.OrderProductResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "404":
          description: Order Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an order
      tags:
      - Orders
  /api/orders/{id}/status:
    put:
      consumes:
      - application/json
      description: Update the status of an existing order
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: Order Status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/courier.UpdateOrderStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.UpdateOrderStatusResponse'
        "400":
          description: Invalid Order ID
          schema:
            type: string
        "404":
          description: Order Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update order status
      tags:
      - Orders
  /api/orders/{orderId}:
    delete:
      description: Remove an order from the system by its ID
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Invalid Order ID
          schema:
            type: string
        "404":
          description: Order Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete an order
      tags:
      - Orders
    get:
      description: Retrieve details of an order by its ID
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.OrderProductResponse'
        "400":
          description: Invalid Order ID
          schema:
            type: string
        "404":
          description: Order Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get order by ID
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Update details of an existing order by its ID
      parameters:
      - description: Order ID
        in: path
        name: orderId
        required: true
        type: string
      - description: Order Data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/courier.UpdateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.OrderResponse'
        "400":
          description: Invalid Order ID
          schema:
            type: string
        "404":
          description: Order Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update order by ID
      tags:
      - Orders
  /api/orders/create:
    post:
      consumes:
      - application/json
      description: Add a new order to the system
      parameters:
      - description: Order Data
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/courier.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/courier.OrderResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new order
      tags:
      - Orders
  /api/orders/history:
    get:
      description: Retrieve the order history for a specific courier
      parameters:
      - description: Courier ID
        in: query
        name: courierId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.GetCourierOrderHistoryResponse'
        "400":
          description: Invalid Courier ID
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get courier order history
      tags:
      - Orders
...
  /api/recommended-orders:
    get:
      description: Retrieve a list of recommended orders for a courier
      parameters:
      - description: Courier ID
        in: query
        name: courierId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.GetRecommendedOrdersResponse'
        "400":
          description: Invalid Courier ID
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get recommended orders for a courier
      tags:
      - Orders
  /api/tasks:
    get:
      description: Retrieve a list of all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.GetAllTasksResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - Tasks
  /api/tasks/{taskId}:
    delete:
      description: Remove a task from the system by its ID
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task deleted successfully
          schema:
            type: string
        "400":
          description: Invalid Task ID
          schema:
            type: string
        "404":
          description: Task Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete task by ID
      tags:
      - Tasks
    get:
      description: Retrieve details of a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.TaskResponse'
        "400":
          description: Invalid Task ID
          schema:
            type: string
        "404":
          description: Task Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Update details of an existing task by its ID
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      - description: Task Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/courier.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/courier.TaskResponse'
        "400":
          description: Invalid Task ID
          schema:
            type: string
        "404":
          description: Task Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update task by ID
      tags:
      - Tasks
  /api/tasks/create:
    post:
      consumes:
      - application/json
      description: Add a new task to the system
      parameters:
      - description: Task Data
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/courier.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/courier.TaskResponse'
        "400":
          description: Invalid Data
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - Tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"